[form-builder] Refactor PTE toolbar to simplify
