+++ packages/@sanity/form-builder/src/inputs/PortableText/Toolbar/DecoratorButtons.tsx
@@ -7,7 +7,6 @@ import FormatCodeIcon from 'part:@sanity/base/format-code-icon'
 import SanityLogoIcon from 'part:@sanity/base/sanity-logo-icon'
 import ToggleButton from 'part:@sanity/components/toggles/button'
 import styles from './DecoratorButtons.css'
-import CollapsibleButtonGroup from './CollapsibleButtonGroup'
 import {
   PortableTextEditor,
   PortableTextFeature,
@@ -24,7 +23,6 @@ type DecoratorItem = PortableTextFeature & {
 type Props = {
   editor: PortableTextEditor
   hotkeys: HotkeyOptions
-  collapsed: boolean
   selection: EditorSelection
 }
 
@@ -46,9 +44,6 @@ function getIcon(type: string, schemaIcon?: Function | string): Function {
 }
 export default class DecoratorButtons extends React.Component<Props, {}> {
   shouldComponentUpdate(nextProps: Props): boolean {
-    if (nextProps.collapsed !== this.props.collapsed) {
-      return true
-    }
     if (nextProps.selection !== this.props.selection) {
       return true
     }
@@ -98,16 +93,8 @@ export default class DecoratorButtons extends React.Component<Props, {}> {
     )
   }
   render(): JSX.Element {
-    const {collapsed} = this.props
     const items = this.getItems()
-    const icon = items[0].icon || (items[0].blockEditor && items[0].blockEditor.icon)
-    if (items.length > 0 && collapsed) {
-      return (
-        <CollapsibleButtonGroup icon={icon || getIcon(items[0].value)}>
-          {items.map(this.renderDecoratorButton)}
-        </CollapsibleButtonGroup>
-      )
-    }
+
     return <div className={styles.root}>{items.map(this.renderDecoratorButton)}</div>
   }
 }
